refactor: ClubApplyPage 전체 구조 리팩토링
