+++ client/src/components/ClubSubmitCard/index.tsx
@@ -1,0 +1,1 @@
+export * from './ClubSubmitCard';
