+++ client/src/pages/ClubApplyPage/DetailQuestionPage/ClubApplyDetailQuestionPage.style.ts
@@ -1,0 +1,7 @@
+import { css } from '@emotion/react';
+
+export const clubApplyDetailQuestionContainer = css`
+    display: flex;
+    flex-direction: column;
+    margin-bottom: -0.5rem;
+`;
