+++ client/src/pages/ClubApplyPage/DetailQuestionPage/ClubApplyDetailQuestionPage.tsx
@@ -1,14 +1,16 @@
 import React from 'react';
-import { clubApplyDetailQuestionContainer, clubApplyFormConatiner } from './ClubApplyPage.style';
 import { Text } from '@components/_common/Text';
 import { TextArea } from '@components/_common/TextArea';
+import { clubApplyDetailQuestionContainer } from './ClubApplyDetailQuestionPage.style';
+import { clubApplyFormContainer } from '../PersonalInfoPage/ClubApplyPersonalInfoPage.style';
+
 interface ClubApplyDetailQuestionPageProps {
     idx: number;
 }
 
 function ClubApplyDetailQuestionPage({ idx }: ClubApplyDetailQuestionPageProps) {
     return (
-        <div css={clubApplyFormConatiner(idx)}>
+        <div css={clubApplyFormContainer(idx)}>
             <div css={clubApplyDetailQuestionContainer}>
                 <Text textAlign="start" type="captionRegular">
                     En#에 지원한 동기가 무엇인가요?
