+++ client/src/pages/ClubApplyPage/DetailQuestionPage/index.tsx
@@ -1,0 +1,1 @@
+export * from './ClubApplyDetailQuestionPage';
