+++ client/src/pages/ClubApplyPage/PersonalInfoPage/index.tsx
@@ -1,0 +1,1 @@
+export * from './ClubApplyPersonalInfoPage';
